Eliminate CodecPlugin dependency in InboundAudioStream,
by separating the codec interface from the plugin interface.
